Add "hasSpecialCloseBehaviour" option to window
This flag will change how the close works.

If the hasSpecialCloseBehaviour is set to true, when the the close
button is clicked, the window will fire a close event and will not close
the window. To close the window it will be necessary to call the "close"
method.

If it is set to false or it is not set it will close normaly.