Added don't executeEventHandlers option && fix restore error
executeEventHandlers option only added, NOT IMPLEMENTED.
This option will be useful when you want the window to emit events
without doing the actual action.
Ie, in virtualization we want to notify that the maximized button is
clicked without maximizing the window. When the virtualized window
answers we tell the ventus window to maximize.