On windows move don't move the content, move a border
This commits changes the behaviour of window movement.
Now, when moving a window instead of moving content on mousemove, we
move only a border. On mouseup we move the content to the desired place.

Because of this change we've had to change some part of the resizing
becuause it was not working properly.